nixos/dovecot: Add proper escaping and support for more value types to plugin settings
The added escaping functions cover the entire range of Dovecot supported
config value constructs other than string-includes and variable references
(neither of which has an equivalent in Nix) and are intended to be reusable
for further structed configuration enhancements. Support for string-includes
and variable references can be re-added (compared to the status quo where they
happen to work due to lack of escaping) when they are needed.